Fix overflow issue in sidebar
* Change values from int to long

+++ src/main/se306group8/scheduleoptimizer/algorithm/RuntimeMonitor.java
@@ -33,15 +33,15 @@ public interface RuntimeMonitor {
 	/** 
 	 * Sets the number of solutions found so far.
 	 **/
-	public void setSchedulesExplored(int number);
+	public void setSchedulesExplored(long number);
 
-	public void setSchedulesInArray(int number);
+	public void setSchedulesInArray(long number);
 	public void setScheduleInArrayStorageSize(int bytes);
 
-	public void setSchedulesInQueue(int number);
+	public void setSchedulesInQueue(long number);
 	public void setScheduleInQueueStorageSize(int bytes);
 
-	public void setSchedulesOnDisk(int number);
+	public void setSchedulesOnDisk(long number);
 	public void setScheduleOnDiskStorageSize(int bytes);
 	
 	default void setScheduleDistribution(int[] distribution, int limit) {  }
