+++ src/main/se306group8/scheduleoptimizer/visualisation/manager/PieChartManager.java
@@ -27,9 +27,9 @@ public class PieChartManager extends Manager {
 
 	@Override
 	protected void updateHook(ObservableRuntimeMonitor monitor) {
-		int schedulesInArray = monitor.getSchedulesInArray();
-		int schedulesInQueue = monitor.getSchedulesInQueue();
-		int schedulesOnDisk = monitor.getSchedulesOnDisk();
+		long schedulesInArray = monitor.getSchedulesInArray();
+		long schedulesInQueue = monitor.getSchedulesInQueue();
+		long schedulesOnDisk = monitor.getSchedulesOnDisk();
 
 		long totalSchedules = schedulesInArray + schedulesInQueue + schedulesOnDisk;
